requirements.txt
requests==2.28.1
python-telegram-bot==13.13
python-dotenv==0.20.0



